Treat all JIT buffer parameters as non-linear parameters
This commit addresses very large kernels that were generated by the
JIT for smaller buffers. This was occuring because of how we were
estimating the JIT parameter sizes.

When targeting linear kernels we only pass the pointer to the kernel
instead of the entire dims and stride arrays. This generates smaller
parameter sets so we calculated kernels based on the pointer size
instead of the Param<T> size.

The issue occured when this kernel came in below the threshold for
evaluation but when it was used as a node to another operation, the
generated kernel became a non-linear kernel requiring us to pass
the Param<T> object for each buffer. This causes the size to go past
the limit of the parameters supported by CUDA and compilation fails.

This commit treats all kernels as non-linear kernel when calculating
the parameter sizes.

Fixes #2436 #2389

(cherry picked from commit 9007742719f6decfb705b49781d6d2aaf23402de)

+++ src/backend/common/jit/Node.hpp
@@ -99,7 +99,7 @@ class Node {
 
     // Return the size of the parameter in bytes that will be passed to the
     // kernel
-    virtual short getParamBytes() const { return 0; }
+    virtual size_t getParamBytes() const { return 0; }
 
     // Return the size of the size of the buffer node in bytes. Zero otherwise
     virtual size_t getBytes() const { return 0; }
